[Decomp][Creature] Cohesion split: extract level/stat scaling to CreatureLevel.cpp
